+++ lib/features/post/presentation/pages/widgets/image_picker_and_anonymous_row.dart
@@ -1,4 +1,4 @@
-part of 'package:sodong_app/features/create_post/presentation/pages/create_post_page.dart';
+part of 'package:sodong_app/features/post/presentation/pages/create_post_page.dart';
 
 class _ImagePickerAndAnonymousRow extends StatelessWidget {
   const _ImagePickerAndAnonymousRow({
