+++ lib/features/post/presentation/pages/widgets/image_preview.dart
@@ -1,4 +1,4 @@
-part of 'package:sodong_app/features/create_post/presentation/pages/create_post_page.dart';
+part of 'package:sodong_app/features/post/presentation/pages/create_post_page.dart';
 
 class ImagePreview extends ConsumerWidget {
   const ImagePreview({super.key, required this.imageFiles});
