+++ lib/features/post/presentation/pages/widgets/top_app_bar_title.dart
@@ -1,4 +1,4 @@
-part of 'package:sodong_app/features/create_post/presentation/pages/create_post_page.dart';
+part of 'package:sodong_app/features/post/presentation/pages/create_post_page.dart';
 
 class _TopAppBarTitle extends StatelessWidget {
   const _TopAppBarTitle();
